+++ docs/release-notes.md
@@ -2,6 +2,12 @@
 ## Possible future features:
 
 
+## 0.0.65
+
+- Remove fields active and inherited for learnsets/testsets/models
+- Remove taxonomy stuff 
+- Add models to get list of learnsets/testsets/models 
+
 ## 0.0.64
 
 - Add models to extended endpoints in summary